+++ query-engine/schema-builder/src/input_types/fields/arguments.rs
@@ -1,5 +1,6 @@
 use super::*;
 use crate::input_types::objects::order_by_objects::OrderByOptions;
+use crate::mutations::create_one;
 use constants::args;
 use datamodel_connector::ConnectorCapability;
 use objects::*;
@@ -23,20 +24,6 @@ pub(crate) fn where_unique_argument(ctx: &mut BuilderContext, model: &ModelRef) 
     }
 }
 
-/// Builds "data" argument intended for the create field.
-/// The data argument is not present if no data can be created.
-pub(crate) fn create_one_arguments(ctx: &mut BuilderContext, model: &ModelRef) -> Option<Vec<InputField>> {
-    let create_types = create_one_objects::create_one_input_types(ctx, model, None);
-    let any_empty = create_types.iter().any(|typ| typ.is_empty());
-    let all_empty = create_types.iter().all(|typ| typ.is_empty());
-
-    if all_empty {
-        None
-    } else {
-        Some(vec![input_field(args::DATA, create_types, None).optional_if(any_empty)])
-    }
-}
-
 /// Builds "where" (unique) argument intended for the delete field.
 pub(crate) fn delete_one_arguments(ctx: &mut BuilderContext, model: &ModelRef) -> Option<Vec<InputField>> {
     where_unique_argument(ctx, model).map(|arg| vec![arg])
@@ -55,7 +42,7 @@ pub(crate) fn update_one_arguments(ctx: &mut BuilderContext, model: &ModelRef) -
 pub(crate) fn upsert_arguments(ctx: &mut BuilderContext, model: &ModelRef) -> Option<Vec<InputField>> {
     where_unique_argument(ctx, model).and_then(|where_unique_arg| {
         let update_types = update_one_objects::update_one_input_types(ctx, model, None);
-        let create_types = create_one_objects::create_one_input_types(ctx, model, None);
+        let create_types = create_one::create_one_input_types(ctx, model, None);
 
         if update_types.iter().all(|typ| typ.is_empty()) || create_types.iter().all(|typ| typ.is_empty()) {
             None
