+++ query-engine/schema-builder/src/input_types/fields/input_fields.rs
@@ -1,5 +1,6 @@
 use super::objects::*;
 use super::*;
+use crate::mutations::create_one;
 use constants::{args, operations};
 use datamodel_connector::ConnectorCapability;
 
@@ -18,8 +19,7 @@ pub(crate) fn filter_input_field(ctx: &mut BuilderContext, field: &ModelField, i
 }
 
 pub(crate) fn nested_create_one_input_field(ctx: &mut BuilderContext, parent_field: &RelationFieldRef) -> InputField {
-    let create_types =
-        create_one_objects::create_one_input_types(ctx, &parent_field.related_model(), Some(parent_field));
+    let create_types = create_one::create_one_input_types(ctx, &parent_field.related_model(), Some(parent_field));
 
     let types: Vec<InputType> = create_types
         .into_iter()
