+++ query-engine/schema-builder/src/input_types/objects/mod.rs
@@ -1,6 +1,5 @@
 pub(super) mod connect_or_create_objects;
 pub(super) mod create_many_objects;
-pub(super) mod create_one_objects;
 pub(super) mod filter_objects;
 pub(super) mod order_by_objects;
 pub(super) mod update_many_objects;
