+++ query-engine/schema-builder/src/input_types/objects/upsert_objects.rs
@@ -1,4 +1,5 @@
 use crate::constants::args;
+use crate::mutations::create_one;
 
 use super::*;
 
@@ -20,7 +21,7 @@ fn nested_upsert_list_input_object(
 ) -> Option<InputObjectTypeWeakRef> {
     let related_model = parent_field.related_model();
     let where_object = filter_objects::where_unique_object_type(ctx, &related_model);
-    let create_types = create_one_objects::create_one_input_types(ctx, &related_model, Some(parent_field));
+    let create_types = create_one::create_one_input_types(ctx, &related_model, Some(parent_field));
     let update_types = update_one_objects::update_one_input_types(ctx, &related_model, Some(parent_field));
 
     if where_object.into_arc().is_empty() || create_types.iter().all(|typ| typ.is_empty()) {
@@ -60,7 +61,7 @@ fn nested_upsert_nonlist_input_object(
     parent_field: &RelationFieldRef,
 ) -> Option<InputObjectTypeWeakRef> {
     let related_model = parent_field.related_model();
-    let create_types = create_one_objects::create_one_input_types(ctx, &related_model, Some(parent_field));
+    let create_types = create_one::create_one_input_types(ctx, &related_model, Some(parent_field));
     let update_types = update_one_objects::update_one_input_types(ctx, &related_model, Some(parent_field));
 
     if create_types.iter().all(|typ| typ.is_empty()) {
